feat(sanitycheck): Rewrite logic to run tests
