+++ src/pvecontrol/sanitycheck/__init__.py
@@ -1,0 +1,1 @@
+from .sanitychecks import SanityCheck, DEFAULT_CHECKS_ORDER
